requirements.txt
numpy==1.14.2
pandas==0.22.0
matplotlib==2.2.0
pyvcf
pysam